Refactor api in the more cleaner routes
